Small cleanup.
